+++ internal/entity/jdkversion.go
@@ -1,0 +1,6 @@
+package entity
+
+type JdkVersion struct {
+	Version string `json:"version"`
+	Url     string `json:"url"`
+}
